BlockCoordinator 구현

+++ Tempus/Domain/Entity/Block/BlockEntity+CoreDataProperties.swift
@@ -16,7 +16,7 @@ extension BlockEntity {
         return NSFetchRequest<BlockEntity>(entityName: "BlockEntity")
     }
 
-    @NSManaged public var divideCount: Int
+    @NSManaged public var divideCount: Int16
     @NSManaged public var createdAt: Double
     @NSManaged public var uuid: UUID
     @NSManaged public var title: String
@@ -24,7 +24,7 @@ extension BlockEntity {
     var toModel: BlockModel {
         return BlockModel(id: uuid,
                           title: title,
-                          divideCount: divideCount)
+                          divideCount: Int(divideCount))
     }
 }
